app/requirements.txt
python-dotenv
requests
beautifulsoup4
pytest
flake8
black